Fix gcc -Wconversion warnings R vectors (incl. NumericVector) are indexed with int, not size_t.

+++ src/Predictor.cpp
@@ -23,15 +23,15 @@
 using std::vector;
 using std::string;
 
-vector<size_t> Predictor::getOrder() const {
-    vector<size_t> index(nTotal);
+vector<int> Predictor::getOrder() const {
+    vector<int> index(nTotal);
     std::iota(index.begin(), index.end(), 0);
     std::sort(index.begin(), index.end(), PredictorComparator(*this));
     return index;
 }
 
-vector<size_t> Predictor::getOrder(const string direction) const {
-    vector<size_t> index(nTotal);
+vector<int> Predictor::getOrder(const string direction) const {
+    vector<int> index(nTotal);
     std::iota(index.begin(), index.end(), 0);
     if (direction == ">") {
       std::sort(index.begin(), index.end(), PredictorComparator(*this));
