+++ src/Predictor.h
@@ -30,27 +30,27 @@
 class Predictor {
   const Rcpp::NumericVector& controls;
   const Rcpp::NumericVector& cases;
-  const size_t nControls, nCases, nTotal;
+  const int nControls, nCases, nTotal;
   public:
     Predictor(const Rcpp::NumericVector& someControls, const Rcpp::NumericVector& someCases): 
               controls(someControls), cases(someCases),
               nControls(someControls.size()), nCases(someCases.size()), nTotal(nControls + nCases) {}
-    double operator[] (const size_t anIdx) const {
+    double operator[] (const int anIdx) const {
       return anIdx < nControls ? controls[anIdx] : cases[anIdx - nControls];
     }
     
-    std::vector<size_t> getOrder() const;
-    std::vector<size_t> getOrder(std::string) const;
+    std::vector<int> getOrder() const;
+    std::vector<int> getOrder(std::string) const;
 
-    bool isControl(const size_t anIdx) const {
+    bool isControl(const int anIdx) const {
       return anIdx < nControls;
     }
 
-    bool isCase(const size_t anIdx) const {
+    bool isCase(const int anIdx) const {
       return anIdx >= nControls;
     }
     
-    bool isValid(const size_t anIdx) const { // Is there a predictor at this index? 
+    bool isValid(const int anIdx) const { // Is there a predictor at this index? 
       return anIdx < nTotal;
     }
     
@@ -62,7 +62,7 @@ class Predictor {
  * PredictorReverseComparator: reverse sort!
 
  * Example usage: 
- * vector<size_t> index(controls.size() + cases.size());
+ * vector<int> index(controls.size() + cases.size());
  * std::iota(index.begin(), index.end(), 0);
  * std::sort(index.begin(), index.end(), ControlCasesComparator(controls, cases));
  */
@@ -70,7 +70,7 @@ class PredictorComparator{
    const Predictor& predictor;
  public:
    PredictorComparator(const Predictor& somePredictor) : predictor(somePredictor) {}
-   bool operator() (size_t i, size_t j) const {
+   bool operator() (int i, int j) const {
      return predictor[i] < predictor[j];
    }
 };
@@ -79,7 +79,7 @@ class PredictorReverseComparator{
    const Predictor& predictor;
  public:
    PredictorReverseComparator(const Predictor& somePredictor) : predictor(somePredictor) {}
-   bool operator() (size_t i, size_t j) const {
+   bool operator() (int i, int j) const {
      return predictor[j] < predictor[i];
    }
 };
